:dizzy: 缓存单测正常执行
