+++ voglander-web/src/main/java/io/github/lunasaw/voglander/web/api/device/DeviceController.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import io.github.lunasaw.voglander.common.constant.ApiConstant;
-import io.github.lunasaw.voglander.manager.assembler.DeviceAssembler;
 import io.github.lunasaw.voglander.manager.domaon.dto.DeviceDTO;
 import io.github.lunasaw.voglander.manager.manager.DeviceManager;
 import io.github.lunasaw.voglander.web.api.device.assembler.DeviceWebAssembler;
